Common.Parsing

Description

Provides functions for parsing String arguments.

Documentation

data AnyParser acc where

GADT that hides the a on Parser a acc. This way we can accept a heterogenous [AnyParser acc] so we can try different parsers at once.

Constructors

AnyParser :: Parser a acc -> AnyParser acc

data Parser a acc

Wraps functions that:

Attempts to parse a String into an a.

Updates acc with a.

That is,

  parseFn :: String -> Maybe a
  updateFn :: acc -> a -> acc

Constructors

ExactParser (String -> Maybe a, acc -> a -> acc)
Parses an exact argument e.g. "-flag".

PrefixParser (String, String -> Maybe a, acc -> a -> acc)
Includes a prefix for parsing e.g. "--arg=val".

data ParseErr

Describes anything other than a successful parse.

Constructors

Err String
Indicates that there was an error when trying to parse String.

Help
Indicates the argument "--help" was passed.

parseAll :: [AnyParser acc] -> [String] -> acc -> Either ParseErr acc

Entrypoint for parsing arguments [String] into acc. Returns

   - Left (Err String): some String could not be parsed by any parser.
   - Left Help: "--help" was found
   - Right acc: every String was parsed successfully by some parser.

In symbols, let P = [AnyParser acc], S = [String] and define \[ p(s) = \begin{cases} 1, &p\mathrm{\,parsed\,}s\,\mathrm{successfully.} \\ 0, &\mathrm{otherwise} \end{cases} \] Then, \[ \mathrm{parseAll}(P)(S) = \mathrm{Right\,acc} \iff \forall s \in S, \exists p \in P \mathrm{\;such\,that\;} p(s) = 1 \]