Core.IO

Description

Exports functions to be used by Core.MonadStaleBranches for IO.

Documentation

errTupleToBranch :: Maybe FilePath -> Text -> ErrOr NameAuthDay -> IO (ErrOr AnyBranch)

Maps a NameAuthDay on FilePath to AnyBranch. Returns Left Err if any errors occur, Right NameAuthDay otherwise.

logIfErr :: forall a. IO a -> IO a

Runs an IO action and logs the error if any occur.

nameToLog :: Maybe FilePath -> ErrOr Name -> IO (ErrOr NameAuthDay)

Retrieves the log information for a given branch Name on FilePath. Returns Left Err if any errors occur, Right NameAuthDay otherwise.

sh :: Text -> Maybe FilePath -> IO Text

Runs a shell command given by Text on FilePath.