Git.FastForward.Types.UpdateResult

Description

Provides functions and types for descrbing the result of an update attempt.

Documentation

data SplitResults

Groups the String Name of all results by UpdateResult.

Constructors

SplitResults

Fields

failures :: [String]

noChanges :: [String]

successes :: [String]

data UpdateResult

Describes the result of an attempt to update a branch.

Constructors

Failure Name
Indicates updating Name failed.

NoChange Name
Indicates updating Name resulted in no change.

Success Name
Indicates updating Name succeeded.

displayResults :: [UpdateResult] -> String

Summarizes [UpdateResult] as a display String.

splitResults :: [UpdateResult] -> SplitResults

Maps [UpdateResult] to intermediate SplitResults.