Git.Stale.Types.Branch

Description

Provides types for holding branch information.

Documentation

data AnyBranch where

Hides BranchStatus type parameter, used for processing Branch a in heterogeneous collections.

Constructors

MergedBranch :: Branch 'Merged -> AnyBranch

UnMergedBranch :: Branch 'UnMerged -> AnyBranch

data BranchStatus where

Promoted data type used for adding type-safe merge status to branches.

Constructors

Merged :: BranchStatus

UnMerged :: BranchStatus

data Branch (a :: BranchStatus) where

Holds branch information with phantom BranchStatus.

Constructors

MkBranch :: Name -> Author -> Day -> Branch a

branchesToName :: Text -> [Branch a] -> Text

Maps each [Branch a] to its Text Name and concatenates results. Attempts to strip out an irrelevant prefix it may have (i.e. remoteName).

mkAnyBranch :: Name -> Author -> Day -> Bool -> AnyBranch

Constructs an AnyBranch wrapper around a Branch a where a is Merged if passed True, UnMerged otherwise.