Git.Stale.Types.Results

Description

Provides a Results type that wraps two maps, one for merged branches and one for unmerged.

Documentation

data Results

Wrapper for holding Map Author [Branch a] data.

Constructors

Results

Fields

mergedMap :: BranchMap 'Merged

Map Author [Branch Merged]

unMergedMap :: BranchMap 'UnMerged

Map Author [Branch UnMerged]

newtype MergedDisp

Constructors

MergedDisp (Text, Int)

newtype UnMergedDisp

Constructors

UnMergedDisp (Text, Int)

newtype ResultsDisp

Constructors

ResultsDisp (MergedDisp, UnMergedDisp)

toResultsDisp :: Text -> Results -> ResultsDisp

Transforms the Results into a ResultsDisp for display purposes. Strips out the prefix from the branches, if it exists.

toResults :: [AnyBranch] -> Results

Maps [AnyBranch] to Results. Pattern matches on AnyBranch to reveal the underlying BranchStatus, ensuring merged and unmerged branches are organized separately.