Parsing.Core

Description

Handles parsing of String args into Env.

Documentation

parseArgs :: Day -> [String] -> Either String Env

Maps Day and parsed [String] args into Right Env, returning any errors as Left String. All arguments are optional (i.e. an empty list is valid), but if any are provided then they must be valid or an error will be returned. Valid arguments are:

  --grep=<string>
      Used for filtering on branch names. Any String is fine, including
      the empty string (i.e. --grep=). Defaults to the empty string.

  --path=<string>
      Path to the git directory. Any String is fine, including the empty
      string (i.e. --path=). Defaults to /share.

  --limit=<days>
      Determines if a branch should be considered stale. Must be a
      non-negative integer. Defaults to 30.

  --branchType=<a[ll]|r[emote]|l[ocal]>
      Determines which branches we should search. Must be one of
      [a, all, r, remote, l, local]. Defaults to remote.

  --remote=<string>
      Name of the remote, used for stripping out the the remote name for
      display purposes. Any String is fine, including the empty string
      (i.e. --remote=). Defaults to origin.

  --master=<string>
      Name of the branch to consider merges against. Any String is fine,
      including the empty string (i.e. --master=). Defaults to origin/master.