Types.Results

Description

Provides a Results type that wraps two maps, one for merged branches and one for unmerged.

Documentation

data Results

Wrapper for holding Map Author [Branch a] data.

Constructors

Results

Fields

mergedMap :: BranchMap 'Merged

Map Author [Branch Merged]

unMergedMap :: BranchMap 'UnMerged

Map Author [Branch UnMerged]

displayResults :: Text -> Results -> Text

Displays Results. Differs from Show in that it is formatted differently and strips the Text prefix from the branch names.

toResults :: [AnyBranch] -> Results

Maps [AnyBranch] to Results. Pattern matches on AnyBranch to reveal the underlying BranchStatus, ensuring merged and unmerged branches are organized separately.